requirements.txt
python-telegram-bot==21.6
playwright==1.49.0
fastapi==0.115.0
uvicorn==0.30.6